fix stuff, some stuff doesn't work
